Portfolio/requirements.txt
# Python packages
Django==5.2.6
Pillow==10.4.0
python-decouple==3.8
whitenoise==6.8.2

# Production packages
gunicorn==23.0.0
psycopg2-binary==2.9.9
dj-database-url==2.3.0

# Development packages (remove in production)
django-debug-toolbar==4.4.6